requirements.txt
langchain
langgraph
langchain_community
dotenv
google-genai
streamlit
pillow
groq